added logout btn to CMS navbar
