+++ components/pages/CMSPages/StasjonerCMSPage.jsx
@@ -2,7 +2,7 @@ import { CMSType } from "@/components/CMSComponents/CMSType"
 
 export const StasjonerCMSPage = () => {
     return (
-        <main className='flex flex-col min-h-screen md:w-80 md:border-r md:border-[#DBDBDB]'>
+        <main className='flex flex-col min-h-[90vh] md:w-80 md:border-r md:border-[#DBDBDB]'>
             <CMSType />
         </main>
     )
